Renames js file
