+++ src/index.js
@@ -1,1 +1,1 @@
-import './js/modal';
+import './js/location-modal';
